+++ src/Response.php
@@ -31,9 +31,9 @@ class Response
     http_response_code(400);
     return ["success" => false, "msg" => $msg];
   }
-  public static function data($data, $msg = "Data Retrieved Successfully", $status = "success")
+  public static function json($data, $msg = "Data Retrieved Successfully", $status = "success")
   {
     http_response_code(200);
-    return ["success" => true, "data" => $data, "msg" => $msg];
+    return ["success" => $status, "data" => $data, "msg" => $msg];
   }
 }