+++ src/Route.php
@@ -86,15 +86,14 @@ class Route
     {
         // on all actions that require a user to be authenticated adding token validation
         // is a smart idea for the purpose of single page applications.
-        if ($this->tokenValidation && $this->howToValidate && !isset($req[0]->headers['Authorization'])) {
+        if ($this->tokenValidation && $this->howToValidate && !isset($req[0]->getHeaders['Authorization'])) {
             echo json_encode(Response::err("Api token missing, please contact your web administrator"));
             exit;
         }
-        if ($this->tokenValidation && $this->howToValidate && !Token::validate($req[0]->headers['Authorization'], $_ENV['KEN_SECRET'])) {
+        if ($this->tokenValidation && $this->howToValidate && !Token::validate($req[0]->getHeaders['Authorization'], $_ENV['KEN_SECRET'])) {
             echo json_encode(Response::err("Invalid token sent to api, error encountered in " . $this->route . " action ,please contact your web administrator"));
             exit;
         }
-        $res = new Response();
-        return call_user_func_array($this->callback, $req, $res);
+        return call_user_func_array($this->callback, $req);
     }
 }